Type-check cleanup: pyright annotations, .pyi stubs, config
Adds pyrightconfig.json and .pyi stubs for C extensions (_lc0_ext,
_features_ext, _mcts_tree). Threads type annotations, cast()s, and
scoped pyright-ignores through the codebase to clear strict-mode
warnings without changing behavior. Hoists module-level mutable init
(e.g. _ADJACENT_FILE_MASKS, _ORIENT_COORDS) into helper functions so
they stop leaking loop variables into module scope.

In selfplay/manager.py: couples batch_process_ply and batch_encode_146
into a single import + logs a warning when the C fast path falls back,
so partial-extension builds fail visibly instead of silently.

+++ chess_anti_engine/model/__init__.py
@@ -15,7 +15,7 @@ class ModelConfig:
     embed_dim: int = 256
     num_layers: int = 6
     num_heads: int = 8
-    ffn_mult: int = 2
+    ffn_mult: float = 2.0
     use_smolgen: bool = True
     use_nla: bool = False
     use_qk_rmsnorm: bool = False
